ecs/requirements.txt
flask
flask_expects_json
mysql-connector-python
zappa